First Commit
Added files from dataanalys leacture.

Tested different ideas for DFS and BFS for solving the return part.

Started an autmation of reading the dataset and creating the inputfile for Vroom

Started an idea to piece several iterations of Vroom input for a final one that would include the return and meeting the capacity constraints and time windows, since the first ideas did not work.